requirements.txt
streamlit
langchain
langchain-google-genai
pypdf
chromadb
python-dotenv